authentication/google/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
PyJWT==2.8.0
sqlalchemy==2.0.20
pydantic==2.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7